ai/my_first_ai_app/requirements.txt
# Flask - Web framework for Python
flask

# OpenAI - Official Python client for the OpenAI API
openai

# python-dotenv - For loading environment variables from .env files
python-dotenv

# gunicorn - Production WSGI HTTP server
gunicorn
